Einsatzleiter/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
defusedxml==0.7.1
Django==5.0.6
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap5==24.2
django-compat==1.0.15
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==24.3
django-htmx==1.18.0
django-tables2==2.7.0
fonttools==4.53.1
fpdf2==2.7.9
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
imap-tools==1.6.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
